refactor: move habit domain related classes to exclusive package
