+++ src/main/java/com/fhsbatista/journal/data/habit/EventRepository.java
@@ -1,7 +1,7 @@
 package com.fhsbatista.journal.data.habit;
 
 import com.fhsbatista.journal.domain.Event;
-import com.fhsbatista.journal.domain.Habit;
+import com.fhsbatista.journal.domain.habit.Habit;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.time.LocalDate;
