+++ src/main/java/com/fhsbatista/journal/data/habit/HabitRepository.java
@@ -1,6 +1,6 @@
 package com.fhsbatista.journal.data.habit;
 
-import com.fhsbatista.journal.domain.Habit;
+import com.fhsbatista.journal.domain.habit.Habit;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface HabitRepository extends JpaRepository<Habit, Long> {
