+++ src/main/java/com/fhsbatista/journal/domain/PerformHabitUsecase.java
@@ -1,8 +1,8 @@
 package com.fhsbatista.journal.domain;
 
 import com.fhsbatista.journal.data.habit.EventRepository;
+import com.fhsbatista.journal.domain.habit.Habit;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.cglib.core.Local;
 import org.springframework.stereotype.Component;
 
 import java.time.LocalDate;
