+++ src/test/java/com/fhsbatista/journal/domain/PerformHabitUsecaseTest.java
@@ -2,6 +2,7 @@ package com.fhsbatista.journal.domain;
 
 import com.fhsbatista.journal.data.habit.EventRepository;
 import com.fhsbatista.journal.domain.area.Area;
+import com.fhsbatista.journal.domain.habit.Habit;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
